moved parse_json out of __init__ and back down to api.py
