+++ sspi_flask_app/api/api.py
@@ -1,9 +1,25 @@
+from flask import Blueprint
+from .. import sspi_raw_api_data
+from bson import json_util
+import json
 
+api_bp = Blueprint(
+    'api_bp', __name__,
+    template_folder='templates',
+    static_folder='static',
+    url_prefix='/api/v1'
+)
 
+# some common utility functions used across the api core functionality
 
-
-def indicator_data_available(IndicatorCode):
+def raw_data_available(IndicatorCode):
     """
     Check if indicator is in database
     """
     return bool(sspi_raw_api_data.find_one({"collection-info.RawDataDestination": IndicatorCode}))
+
+def parse_json(data):
+    return json.loads(json_util.dumps(data))
+
+def print_json(data):
+    print(json.dumps(data, indent=4, sort_keys=True))
